+++ src/bookmarks/utils.js
@@ -7,11 +7,6 @@ export const getCurrentProfileBookmarks = async (pathToBookmarks) => {
   try {
     const currentBookmarksFileData = await readFile(pathToBookmarks, { encoding: 'utf-8' });
     bookmarks = JSON.parse(currentBookmarksFileData);
-    if (bookmarks.bookmark_bar) {
-      return bookmarks;
-    }
-
-    return bookmarks.roots || {};
   } catch (error) {
     console.log(error);
   }